[REFACTOR] fix dir structure
