+++ 1
@@ -1,0 +1,6 @@
+Merge branch 'dev' of https://github.com/TBD-capstone/UNI-frontend into 66-fetch-axios
+# Please enter a commit message to explain why this merge is necessary,
+# especially if it merges an updated upstream into a topic branch.
+#
+# Lines starting with '#' will be ignored, and an empty message aborts
+# the commit.
