+++ uni_frontend/src/pages/login-page/login.css
@@ -18,7 +18,7 @@
     height: 20vw;
     max-width: 120px;
     max-height: 120px;
-    background-image: url('../public/UNI_Logo.png');
+    background-image: url('../../../public/UNI_Logo.png');
     background-size: contain;
     background-repeat: no-repeat;
     margin-bottom: 20px;
