+++ uni_frontend/src/pages/style/index.css
@@ -1,12 +1,12 @@
 @font-face {
   font-family: "GmarketSansTTFMedium";
   font-weight: normal;
-  src: url("./fonts/GmarketSansTTFMedium.ttf") format("truetype");
+  src: url("../../fonts/GmarketSansTTFMedium.ttf") format("truetype");
 }
 @font-face {
   font-family: "NotoSansSC-VariableFont_wght";
   font-weight: normal;
-  src: url("./fonts/NotoSansSC-VariableFont_wght.ttf") format("truetype");
+  src: url("../../fonts/NotoSansSC-VariableFont_wght.ttf") format("truetype");
 }
 
 body {
